Use env_check to check model types
